Add missing maps

+++ Persistence/database/accounts_ships/structure.sql
@@ -4,20 +4,20 @@
 --
 CREATE TABLE `accounts_ships`
 (
-    `id`          int     NOT NULL AUTO_INCREMENT
+    `id`          int      NOT NULL AUTO_INCREMENT
         COMMENT 'Primary Key.',
-    `accounts_id` int     NOT NULL,
-    `ships_id`    tinyint NOT NULL,
-    `maps_id`     tinyint NOT NULL,
-    `position`    bigint  NOT NULL DEFAULT 0
+    `accounts_id` int      NOT NULL,
+    `ships_id`    tinyint  NOT NULL,
+    `maps_id`     smallint NOT NULL,
+    `position`    bigint   NOT NULL DEFAULT 0
         COMMENT 'Position on map.',
-    `health`      int     NOT NULL DEFAULT 0
+    `health`      int      NOT NULL DEFAULT 0
         COMMENT 'Current health points',
-    `shield`      int     NOT NULL DEFAULT 0
+    `shield`      int      NOT NULL DEFAULT 0
         COMMENT 'Current shield points',
-    `nanohull`    int     NOT NULL DEFAULT 0
+    `nanohull`    int      NOT NULL DEFAULT 0
         COMMENT 'Current nanohull points',
-    `gfx`         tinyint NOT NULL DEFAULT 0
+    `gfx`         tinyint  NOT NULL DEFAULT 0
         COMMENT 'Ship graphic (for WIZ-X).',
 
     CONSTRAINT `accounts_ships_pk` PRIMARY KEY (`id`)
