+++ Persistence/database/maps_stations/structure.sql
@@ -4,12 +4,12 @@
 --
 CREATE TABLE `maps_stations`
 (
-    `id`          tinyint NOT NULL AUTO_INCREMENT
+    `id`          tinyint  NOT NULL AUTO_INCREMENT
         COMMENT 'Primary Key.',
-    `position`    bigint  NOT NULL DEFAULT 0
+    `position`    bigint   NOT NULL DEFAULT 0
         COMMENT 'Position on map.',
-    `maps_id`     tinyint NULL     DEFAULT NULL,
-    `factions_id` tinyint NULL     DEFAULT NULL,
+    `maps_id`     smallint NULL     DEFAULT NULL,
+    `factions_id` tinyint  NULL     DEFAULT NULL,
 
     CONSTRAINT `maps_stations_pk` PRIMARY KEY (`id`)
 ) ENGINE InnoDB
