+++ Persistence/src/main/java/com/kalaazu/persistence/entity/MapsNpcsEntity.java
@@ -18,7 +18,7 @@ public class MapsNpcsEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id", nullable = false)
-    private byte id;
+    private short id;
 
     @Basic
     @Column(name = "amount", nullable = false)
@@ -30,7 +30,7 @@ public class MapsNpcsEntity {
 
     @Basic
     @Column(name= "maps_id", nullable = false, insertable = false, updatable = false)
-    private byte mapsId = 0;
+    private short mapsId = 0;
 
     @ManyToOne(fetch = FetchType.EAGER)
     @JoinColumn(name = "npcs_id", referencedColumnName = "id", nullable = false)
